new link state folder
