+++ LoZGame/LoZGame.csproj
@@ -89,9 +89,9 @@
     <Compile Include="Managers\EntityManagers\ProjectileManager.cs" />
     <Compile Include="Managers\ItemManager.cs" />
     <Compile Include="Managers\BlockManager.cs" />
-    <Compile Include="Player\PickupItemState.cs" />
-    <Compile Include="Player\UseItemState.cs" />
-    <Compile Include="Player\DieState.cs" />
+    <Compile Include="Player\PlayerStates\PickupItemState.cs" />
+    <Compile Include="Player\PlayerStates\UseItemState.cs" />
+    <Compile Include="Player\PlayerStates\DieState.cs" />
     <Compile Include="Sprites\ItemSpriteClasses\Arrow.cs" />
     <Compile Include="Sprites\ItemSpriteClasses\BlueCandle.cs" />
     <Compile Include="Sprites\ItemSpriteClasses\BlueRing.cs" />
@@ -159,13 +159,13 @@
     <Compile Include="Interfaces\IPlayerState.cs" />
     <Compile Include="Interfaces\IPlayer.cs" />
     <Compile Include="Interfaces\ISprite.cs" />
-    <Compile Include="Player\AttackState.cs" />
-    <Compile Include="Player\NullState.cs" />
-    <Compile Include="Player\MoveRightState.cs" />
-    <Compile Include="Player\MoveLeftState.cs" />
-    <Compile Include="Player\MoveDownState.cs" />
-    <Compile Include="Player\MoveUpState.cs" />
-    <Compile Include="Player\IdleState.cs" />
+    <Compile Include="Player\PlayerStates\AttackState.cs" />
+    <Compile Include="Player\PlayerStates\NullState.cs" />
+    <Compile Include="Player\PlayerStates\MoveRightState.cs" />
+    <Compile Include="Player\PlayerStates\MoveLeftState.cs" />
+    <Compile Include="Player\PlayerStates\MoveDownState.cs" />
+    <Compile Include="Player\PlayerStates\MoveUpState.cs" />
+    <Compile Include="Player\PlayerStates\IdleState.cs" />
     <Compile Include="Player\Link.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
